Address review comments. Disable tests properly.
