+++ sycl/test/spec_const/spec_const_hw.cpp
@@ -1,9 +1,11 @@
 // RUN: %clangxx -fsycl %s -o %t.out
 // RUN: env SYCL_DEVICE_TYPE=HOST %t.out
-// TODO: re-enable after OpenCL RT is fixed:
-// RUNx: %CPU_RUN_PLACEHOLDER %t.out
-// RUNx: %GPU_RUN_PLACEHOLDER %t.out
-// RUNx: %ACC_RUN_PLACEHOLDER %t.out
+// RUN: %CPU_RUN_PLACEHOLDER %t.out
+// RUN: %GPU_RUN_PLACEHOLDER %t.out
+// RUN: %ACC_RUN_PLACEHOLDER %t.out
+// TODO: re-enable after CI drivers are updated to newer which support spec
+// constants:
+// XFAIL: acc,cpu,cuda,gen
 //
 //==----------- spec_const_hw.cpp ------------------------------------------==//
 //
