+++ sycl/test/spec_const/spec_const_redefine.cpp
@@ -1,11 +1,11 @@
 // RUN: %clangxx -fsycl %s -o %t.out
 // RUN: env SYCL_DEVICE_TYPE=HOST %t.out
-// TODO re-enable CPU device test once CI CPU OCL RT is updated:
-// RUNx: env SYCL_PI_TRACE=2 %CPU_RUN_PLACEHOLDER %t.out 2>&1 %CPU_CHECK_PLACEHOLDER
+// RUN: env SYCL_PI_TRACE=2 %CPU_RUN_PLACEHOLDER %t.out 2>&1 %CPU_CHECK_PLACEHOLDER
 // RUN: env SYCL_PI_TRACE=2 %GPU_RUN_PLACEHOLDER %t.out 2>&1 %GPU_CHECK_PLACEHOLDER
 // RUN: env SYCL_PI_TRACE=2 %ACC_RUN_PLACEHOLDER %t.out 2>&1 %ACC_CHECK_PLACEHOLDER
-// TODO the test currently fails on these devices:
-// XFAIL: acc,cuda
+// TODO: re-enable after CI drivers are updated to newer which support spec
+// constants:
+// XFAIL: acc,cpu,cuda,gen
 //
 //==----------- spec_const_redefine.cpp ------------------------------------==//
 //
